requirements.txt
crewai
langchain-google-genai #gemini model
load_dotenv
crewai_tools
streamlit